+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Change Log
 
-## [5.46.1](https://github.com/plivo/plivo-dotnet/tree/v5.46.1) (2024-05-28)
+## [5.47.1](https://github.com/plivo/plivo-dotnet/tree/v5.46.1) (2024-05-28)
 **Feature - Adding filter support for List Application API**
 - Added new filer param `appName` in list application API
+
+## [5.47.0](https://github.com/plivo/plivo-dotnet/tree/v5.47.0) (2024-05-28)
+**Feature - Number Masking Feature Added**
+- Number Masking APIs added to create, update, delete and list sessions
 
 ## [5.46.0](https://github.com/plivo/plivo-dotnet/tree/v5.46.0) (2024-05-20)
 **Feature - Adding support for location whatsapp messages**